backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
python-multipart>=0.0.17
psycopg2-binary>=2.9.9  # PostgreSQL adapter for Python

# AI/LLM dependencies
openai>=1.10.0  # For both DeepSeek (OpenAI-compatible) and OpenAI embeddings
chromadb>=0.4.22  # Vector database
tiktoken>=0.5.2  # Token counting
redis>=5.0.0
rq>=1.16.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
